# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Book sales platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Validation error
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
  /payments:
    post:
      operationId: initiatePayment
      tags: [payments]
      summary: Initiate a payment for an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Payment initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "400":
          description: Validation error
  /payments/{id}:
    get:
      operationId: getPayment
      tags: [payments]
      summary: Get payment status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "404":
          description: Not found
  /payments/{id}/refresh:
    post:
      operationId: refreshPayment
      tags: [payments]
      summary: Refresh payment status from the gateway
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "404":
          description: Not found
  /payments/{id}/simulate:
    patch:
      operationId: simulatePayment
      tags: [payments]
      summary: Simulate payment outcome (success/failure)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentSimulate"
      responses:
        "200":
          description: Payment updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "404":
          description: Not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    OrderInput:
      type: object
      required: [fullName, phone, email, deliveryAddress, city, productType, quantity]
      properties:
        fullName:
          type: string
          minLength: 2
        phone:
          type: string
          minLength: 8
        email:
          type: string
          format: email
        deliveryAddress:
          type: string
          minLength: 5
        city:
          type: string
          minLength: 2
        productType:
          type: string
          enum: [paperback, hardcover]
        quantity:
          type: integer
          minimum: 1
        notes:
          type: ["string", "null"]
    Order:
      type: object
      required: [id, fullName, phone, email, deliveryAddress, city, productType, quantity, totalAmount, status, createdAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        phone:
          type: string
        email:
          type: string
        deliveryAddress:
          type: string
        city:
          type: string
        productType:
          type: string
        quantity:
          type: integer
        totalAmount:
          type: number
        status:
          type: string
          enum: [pending, confirmed, awaiting_delivery, picked_up, in_transit, delivered, failed_delivery, cancelled]
        courierId:
          type: ["integer", "null"]
        trackingNotes:
          type: ["string", "null"]
        deliveryPaymentMethod:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        payment:
          $ref: "#/components/schemas/Payment"
    PaymentInput:
      type: object
      required: [orderId, method]
      properties:
        orderId:
          type: integer
        method:
          type: string
          enum: [airtel_money, mtn_money, zamtel_money, visa_mastercard, bank_transfer, mobile_money_on_delivery]
        phoneNumber:
          type: ["string", "null"]
        accountName:
          type: ["string", "null"]
    Payment:
      type: object
      required: [id, orderId, method, amount, status, reference, createdAt]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        method:
          type: string
        amount:
          type: number
        status:
          type: string
          enum: [pending, successful, failed]
        reference:
          type: string
        phoneNumber:
          type: ["string", "null"]
        accountName:
          type: ["string", "null"]
        gatewayTransactionId:
          type: ["string", "null"]
        gatewayStatus:
          type: ["string", "null"]
        gatewayCheckoutUrl:
          type: ["string", "null"]
        gatewayProviderReference:
          type: ["string", "null"]
        gatewayIdempotencyKey:
          type: ["string", "null"]
        gatewayLastPayload:
          type: ["object", "null"]
          additionalProperties: true
        lastCheckedAt:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        expiresAt:
          type: ["string", "null"]
        failureReason:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    PaymentSimulate:
      type: object
      required: [outcome]
      properties:
        outcome:
          type: string
          enum: [successful, failed]
...